New DEBUG Schema that uses staging for testing
We added a new Schema to the app for DEBUG, if DEBUG is set in xCode we use the https://bramptonsoccer.flywheel.com environment for testing. HTTP AUTH must be disabled for this to work.  We've also now adjusted all the API endpoints/URLs to use whatever the Schema dictates to access the WP sites.